Move flags and ensure enabled for specific options
